test: relocate inbound test files to sibling __tests__ for coverage check
Move ci-completion-checker.test.ts and inbound-job-handler.test.ts (the
latter renamed to job-handler.test.ts to mirror its production file)
from packages/server/src/services/__tests__/ to
packages/server/src/services/inbound/__tests__/. The preflight coverage
check (.claude/skills/orchestrator/preflight-check.js) expects test
files to live in a sibling __tests__/ directory of the production file,
not a grandparent's. The previous location worked at runtime but
failed coverage-check whenever an inbound/ production file was modified.
